fixed pointer casting which fails on 64bit systems

+++ drm_slp/drm_slp_bufmgr.c
@@ -82,7 +82,7 @@ unsigned int
 drm_slp_bo_map(drm_slp_bo bo, int device, int opt)
 {
     tbm_bo_handle bo_handle;
-    unsigned int ret = 0;
+    unsigned long ret = 0;
 
     bo_handle = tbm_bo_map ((tbm_bo)bo, device, opt);
     if (bo_handle.ptr == NULL)
@@ -97,10 +97,10 @@ drm_slp_bo_map(drm_slp_bo bo, int device, int opt)
         case TBM_DEVICE_2D:
         case TBM_DEVICE_3D:
         case TBM_DEVICE_MM:
-            ret = (unsigned int)bo_handle.u32;
+            ret = (unsigned long)bo_handle.u32;
             break;
         case TBM_DEVICE_CPU:
-            ret = (unsigned int)bo_handle.ptr;
+            ret = (unsigned long)bo_handle.ptr;
             break;
         default:
             fprintf (stderr, "[libdrm_slp:%d]: error wrong device type\n", getpid());
